requirements.txt
openai==1.52.0
spacy==3.8.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
requests==2.32.3
